[refactoring] DevPage
