+++ src/pages/DevPage/DevPage.ts
@@ -3,17 +3,6 @@ function buildSettingsPage() {
   return html;
 }
 
-export function demoModals() {
-  const btnOpenModal = document.getElementById("openModalBtn");
-  const modal: any = document.getElementById('modal');
-  btnOpenModal?.addEventListener('click', () => {
-    console.log('btnOpenModal')
-    if (modal) {
-      modal.open();
-    }
-  })
-}
-
 export function DevPage() {
   return buildSettingsPage();
 }
